fix(core): moved parserAndSerializerDynamicModifiers from core
