+++ src/core/ExtensionsManager.ts
@@ -1,17 +1,14 @@
 import MarkdownIt, {PresetName} from 'markdown-it';
-import type Token from 'markdown-it/lib/token';
-import {Node, Schema} from 'prosemirror-model';
+import {Schema} from 'prosemirror-model';
 import type {Plugin} from 'prosemirror-state';
-import {v5} from 'uuid';
 
 import {ActionsManager} from './ActionsManager';
 import {ExtensionBuilder} from './ExtensionBuilder';
 import {ParserTokensRegistry} from './ParserTokensRegistry';
 import {SchemaSpecRegistry} from './SchemaSpecRegistry';
 import {SerializerTokensRegistry} from './SerializerTokensRegistry';
-import {MarkdownParserDynamicModifier, TokenAttrs} from './markdown/MarkdownParser';
+import {MarkdownParserDynamicModifier} from './markdown/MarkdownParser';
 import {MarkdownSerializerDynamicModifier} from './markdown/MarkdownSerializerDynamicModifier';
-import {MarkupManager} from './markdown/MarkupManager';
 import {TransformFn} from './markdown/ProseMirrorTransformer';
 import type {ActionSpec} from './types/actions';
 import type {
@@ -22,7 +19,6 @@ import type {
     ExtensionSpec,
 } from './types/extension';
 import type {MarkViewConstructor, NodeViewConstructor} from './types/node-views';
-import {SerializerNodeToken, SerializerState} from './types/serializer';
 
 type ExtensionsManagerParams = {
     extensions: Extension;
@@ -85,13 +81,8 @@ export class ExtensionsManager {
         }
 
         if (options.allowDynamicModifiers) {
-            const markupManager = new MarkupManager();
-            this.#parserDynamicModifier =
-                options?.dynamicModifiers?.parser ??
-                this.createParserDynamicModifier(markupManager);
-            this.#serializerDynamicModifier =
-                options?.dynamicModifiers?.serializer ??
-                this.createSerializerDynamicModifier(markupManager);
+            this.#parserDynamicModifier = options?.dynamicModifiers?.parser;
+            this.#serializerDynamicModifier = options?.dynamicModifiers?.serializer;
         }
 
         // TODO: add prefilled context
@@ -186,100 +177,4 @@ export class ExtensionsManager {
             this.#markViews[name] = view(this.#deps);
         }
     }
-
-    private createParserDynamicModifier(markupManager: MarkupManager) {
-        return new MarkdownParserDynamicModifier(createParserDynamicModifierConfig(markupManager));
-    }
-
-    private createSerializerDynamicModifier(markupManager: MarkupManager) {
-        return new MarkdownSerializerDynamicModifier(
-            createSerializerDynamicModifierConfig(markupManager),
-        );
-    }
-}
-
-const YFM_TABLE_TOKEN_ATTR = 'data-token-id';
-const YFM_TABLE_NODE_ATTR = 'data-node-id';
-const PARENTS_WITH_AFFECT = ['blockquote', 'yfm_tabs'];
-
-/**
- * Creates a hook for injecting custom logic into the parsing process via `MarkdownParserDynamicModifier`,
- * allowing extensions beyond the fixed parsing rules defined by the schema.
- *
- * Dynamically configures parsing for `yfm_table` elements:
- * - Assigns a unique `data-token-id` to each token.
- * - Captures and stores the raw Markdown using `MarkupManager`.
- * - Links the token to its corresponding node via `data-node-id`.
- * - Adds the `data-node-id` attribute to the list of allowed attributes.
- */
-function createParserDynamicModifierConfig(markupManager: MarkupManager) {
-    return {
-        ['yfm_table']: {
-            processToken: [
-                (token: Token, _: number, rawMarkup: string) => {
-                    const {map} = token;
-
-                    if (map) {
-                        const content = rawMarkup.split('\n').slice(map[0], map[1]).join('\n');
-                        const tokenId = v5(content, markupManager.getNamespace());
-
-                        token.attrSet(YFM_TABLE_TOKEN_ATTR, tokenId);
-                        markupManager.setMarkup(tokenId, content);
-                    }
-                    return token;
-                },
-            ],
-            processNodeAttrs: [
-                (token: Token, attrs: TokenAttrs) => ({
-                    ...attrs,
-                    [YFM_TABLE_NODE_ATTR]: token.attrGet(YFM_TABLE_TOKEN_ATTR),
-                }),
-            ],
-            processNode: [
-                (node: Node) => {
-                    const nodeId = node.attrs[YFM_TABLE_NODE_ATTR];
-                    if (nodeId) {
-                        markupManager.setNode(nodeId, node);
-                    }
-                    return node;
-                },
-            ],
-            allowedAttrs: [YFM_TABLE_NODE_ATTR],
-        },
-    };
-}
-
-/**
- * Creates a hook for injecting custom logic into the serialization process via `MarkdownSerializerDynamicModifier`,
- * allowing extensions beyond the standard serialization rules defined by the schema.
- *
- * Dynamically configures serialization for `yfm_table` elements:
- * - Retrieves the original Markdown using the `data-node-id` attribute.
- * - Uses the original Markdown if the node matches the saved version.
- * - Falls back to schema-based rendering if the node structure, attributes, or parent elements affect it.
- */
-function createSerializerDynamicModifierConfig(markupManager: MarkupManager) {
-    return {
-        ['yfm_table']: {
-            processNode: [
-                (
-                    state: SerializerState,
-                    node: Node,
-                    parent: Node,
-                    index: number,
-                    callback?: SerializerNodeToken,
-                ) => {
-                    const nodeId = node.attrs[YFM_TABLE_NODE_ATTR];
-                    const savedNode = markupManager.getNode(nodeId);
-
-                    if (!PARENTS_WITH_AFFECT.includes(parent?.type?.name) && savedNode?.eq(node)) {
-                        state.write(markupManager.getMarkup(nodeId) + '\n');
-                        return;
-                    }
-
-                    callback?.(state, node, parent, index);
-                },
-            ],
-        },
-    };
 }
